feat: image_sequence hold_fraction timing redesign + element snap merge (#63)
* fix: merge element-derived snap positions into the rendered slide

Wire the existing RenderedElement.snap_positions channel through the slide renderer into SlideHTML, so image-sequence frame snaps (already reported by 'scrolly introspect snaps' as the merged set) actually reach the built deck. Widen the rendered snap list to float. Previously an image-sequence slide with no explicit snap_positions snapped only to its endpoints.


* feat: redesign image_sequence timing around hold_fraction and frame-grid snaps

Replace the absolute hold field with hold_fraction (fraction of frame_distance, default 0.2). Frames now snap to the frame grid (scroll_offset + i*frame_distance) with a symmetric hold, and fade_in/fade_out bound the timeline exactly ([scroll_offset - fade_in, last_snap + fade_out]) — removing the hold/2 fade-out overshoot.

Migrates the worked-example and stacked-diffs decks: the hero slides drop their hand-mirrored frame snaps (now merged automatically) and shift scroll_offset by the old hold/2 so the grid snaps land on the old hold-centres, keeping the step/banner/pan keyframes aligned. The worked-example cast/investigation sequences omit hold_fraction to exercise the default.

BREAKING CHANGE: image_sequence 'hold' is removed; use 'hold_fraction'.


* chore: add explicit start/end snap stops to worked-example image sequences

The four image-sequence slides gained per-frame snaps from the merge but lacked the start (0) and end (scroll_range) settle points; add them as author snap_positions so the slide start and end are snappable, merged with the frame snaps.


* fix: align worked-example reveal-slide snaps with their content beats

maint-bg, pressure, and handover had evenly-spaced author snaps that landed between their opacity-keyframe reveals; retarget them to each element's settle point (plus start/end) so scrolling snaps onto freshly-revealed content.


* fix: drop build-trojan's dead scroll tail

Its only animation is the mermaid fade-in over [0, 400]; the [400, 800] tail was static (snaps at 400 and 800 showed the same frame). Shorten scroll_range to 400 so the fade spans the whole slide, with snaps at start and end.


* fix: drop anomaly's dead scroll tail

Its only animation is the pull-quote fade over [0, 500]; the [500, 1000] tail was static. Shorten scroll_range to 500 so the fade spans the slide, snaps [0, 500] — same treatment as build-trojan.


---------

+++ scrolly/slide/html.py
@@ -22,7 +22,7 @@ class SlideHTML:
     scroll_speed: float = 1.0
     font_scale: float = 1.0
     assets: tuple[Path, ...] = ()
-    snap_positions: tuple[int, ...] = ()
+    snap_positions: tuple[float, ...] = ()
     reverse: bool = False
     has_mermaid: bool = False
